pyproject.toml
[project]
name = "orders-service"
version = "1.0.0"
description = ""
authors = [
    {name = "AjosephRR",email = "ajrr9866@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic (>=2.13.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "orders_service", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "ruff (>=0.15.10,<0.16.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "mypy (>=1.20.1,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pip-audit (>=2.10.0,<3.0.0)",
    "safety (>=3.7.0,<4.0.0)"
]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = ["pytest-cache-files-*"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.black]
line-length = 88
target-version = ["py312"]
extend-exclude = "pytest-cache-files-.*"

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
pythonpath = ["src"]
testpaths = ["tests"]
